+++ apps/ui-dev/pages/ui-component/visual-bell-target.jsx
@@ -1,5 +1,5 @@
 import { start, useCallback, useState } from "@elf-framework/sapa";
-import { bell, Button } from "@elf-framework/ui";
+import { toast, Button } from "@elf-framework/ui";
 import "@elf-framework/ui/style.css";
 
 import { UILayout } from "../../layouts/UILayout";
@@ -8,7 +8,7 @@ start(function () {
   const [direction, setDirection] = useState("top-left");
 
   const nextFunc = useCallback(() => {
-    const visualBell = bell({
+    const visualBell = toast({
       content: direction,
       direction,
       delay: 2000,